requirement.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==3.0.1
attrs==25.1.0
beautifulsoup4==4.13.4
black==25.1.0
blockbuster==1.5.24
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
coverage==7.8.0
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-rs==0.29.1
langchain==0.3.24
langchain-core==0.3.59
langchain-deepseek==0.1.3
langchain-google-genai==2.0.10
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-api==0.2.21
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.11
langgraph-sdk==0.1.66
langsmith==0.3.42
loguru==0.7.2
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.1.0
openai==1.76.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pyjwt==2.10.1
pyparsing==3.2.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
